Update the bindings to not callback directly into the OCaml runtime
Instead, we signal from the GCD callbacks to a condition variable
that unblocks an OCaml thread.  This is because there is no 1:1
mapping between GCD queues and pthreads, and so registering each
of them with the OCaml GC is problematic.  With this solution, the
only communication back to the OCaml GC is via a single condition
variable.