Add support for reusing interfaces/uuids
The UUID can be reused between calls to Vmnet to get an interface with the same
IP. This change adds an optional UUID parameter to init, as well as helper
functions for reading the UUID from an existing interface t. If UUID is nil or
not specified it will be auto generated as before, so this should be a
non-breaking change.